Zmiana typu wartości położenia slidera (w niektórych przeglądarkach mobilnym wartośc położenia jest typu double, nie int + dodanie informacji do pliku README jak włączyć dostęp aplikacji do automatycznego wysyłania maili, jeżeli Google wyłączy ten dostęp)

+++ src/main/java/com/pokemoncards/controller/SelectOptionController.java
@@ -26,7 +26,7 @@ public class SelectOptionController {
 
 	@PostMapping(value = "/scroll")
 	@ResponseStatus(value = HttpStatus.OK)
-	private void scrollPosition(@ModelAttribute(value = "scroll") int scroll) {
+	private void scrollPosition(@ModelAttribute(value = "scroll") double scroll) {
 		sessionData.setScrollPosition(scroll);
 	}
